Reorganize CLI commands.
