+++ quickwit-cli/tests/cli.rs
@@ -23,11 +23,12 @@ mod helpers;
 
 use std::io::Read;
 use std::path::Path;
+use std::str::from_utf8;
 
 use anyhow::Result;
 use helpers::{TestEnv, TestStorageType};
 use predicates::prelude::*;
-use quickwit_cli::{create_index_cli, CreateIndexArgs};
+use quickwit_cli::index::{create_index_cli, CreateIndexArgs};
 use quickwit_common::rand::append_random_suffix;
 use quickwit_metastore::{Metastore, MetastoreUriResolver};
 use serde_json::{Number, Value};
@@ -39,9 +40,9 @@ use crate::helpers::{create_test_env, make_command, spawn_command};
 fn create_logs_index(test_env: &TestEnv, index_id: &str) {
     make_command(
         format!(
-            "new --index-uri {} --index-id {} --index-config-path {} --metastore-uri {}",
-            test_env.index_uri(index_id),
+            "index create {} --index-uri {} --index-config-path {} --metastore-uri {}",
             index_id,
+            test_env.index_uri(index_id),
             test_env.resource_files["config"].display(),
             test_env.metastore_uri,
         )
@@ -54,7 +55,7 @@ fn create_logs_index(test_env: &TestEnv, index_id: &str) {
 fn index_data(index_id: &str, input_path: &Path, metastore_uri: &str, data_dir_path: &Path) {
     make_command(
         format!(
-            "index --index-id {} --input-path {} --metastore-uri {} --data-dir-path {}",
+            "index index {} --input-path {} --metastore-uri {} --data-dir-path {}",
             index_id,
             input_path.display(),
             metastore_uri,
@@ -90,7 +91,7 @@ async fn test_cmd_new() -> Result<()> {
     assert_eq!(index_metadata.is_ok(), true);
 
     // Attempt to create with ill-formed new command.
-    make_command("new")
+    make_command("index create")
         .assert()
         .failure()
         .stderr(predicate::str::contains("--index-uri <INDEX URI>"));
@@ -106,9 +107,9 @@ fn test_cmd_new_on_existing_index() -> Result<()> {
 
     make_command(
         format!(
-            "new --index-uri {} --index-id {} --index-config-path {} --metastore-uri {}",
-            test_env.index_uri(&index_id),
+            "index create {} --index-uri {} --index-config-path {} --metastore-uri {}",
             index_id,
+            test_env.index_uri(&index_id),
             test_env.resource_files["config"].display(),
             test_env.metastore_uri,
         )
@@ -126,7 +127,7 @@ fn test_cmd_index_on_non_existing_index() -> Result<()> {
     let test_env = create_test_env(TestStorageType::LocalFileSystem)?;
     make_command(
         format!(
-            "index --index-id {} --input-path {} --metastore-uri {} --data-dir-path {}",
+            "index index {} --input-path {} --metastore-uri {} --data-dir-path {}",
             "index-does-no-exist",
             test_env.resource_files["logs"].display(),
             test_env.metastore_uri,
@@ -148,7 +149,7 @@ fn test_cmd_index_on_non_existing_file() -> Result<()> {
     create_logs_index(&test_env, &index_id);
     make_command(
         format!(
-            "index --index-id {} --input-path {} --metastore-uri {} --data-dir-path {}",
+            "index index {} --input-path {} --metastore-uri {} --data-dir-path {}",
             index_id,
             test_env
                 .data_dir_path
@@ -182,7 +183,7 @@ fn test_cmd_index_simple() -> Result<()> {
     let log_path = test_env.resource_files["logs"].clone();
     make_command(
         format!(
-            "index --index-id {} --metastore-uri {} --data-dir-path {}",
+            "index index {} --metastore-uri {} --data-dir-path {}",
             index_id,
             test_env.metastore_uri,
             test_env.data_dir_path.display()
@@ -215,14 +216,15 @@ fn test_cmd_search() -> Result<()> {
 
     make_command(
         format!(
-            "search --metastore-uri {} --index-id {} --query level:info",
-            test_env.metastore_uri, index_id,
+            "index search {} --metastore-uri {} --query level:info",
+            index_id, test_env.metastore_uri,
         )
         .as_str(),
     )
     .assert()
     .success()
     .stdout(predicate::function(|output: &[u8]| {
+        println!("{}", from_utf8(output).unwrap());
         let result: Value = serde_json::from_slice(output).unwrap();
         result["numHits"] == Value::Number(Number::from(2i64))
     }));
@@ -230,9 +232,8 @@ fn test_cmd_search() -> Result<()> {
     // search with tags
     make_command(
         format!(
-            "search --metastore-uri {} --index-id {} --query level:info --tags city:paris \
-             device:rpi",
-            test_env.metastore_uri, index_id,
+            "index search {} --metastore-uri {} --query level:info --tags city:paris device:rpi",
+            index_id, test_env.metastore_uri,
         )
         .as_str(),
     )
@@ -245,8 +246,8 @@ fn test_cmd_search() -> Result<()> {
 
     make_command(
         format!(
-            "search --metastore-uri {} --index-id {} --query level:info --tags city:conakry",
-            &test_env.metastore_uri, index_id,
+            "index search {} --metastore-uri {} --query level:info --tags city:conakry",
+            index_id, &test_env.metastore_uri,
         )
         .as_str(),
     )
@@ -268,7 +269,7 @@ fn test_cmd_delete_index_dry_run() -> Result<()> {
     // Empty index.
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {} --dry-run",
+            "index delete {} --metastore-uri {} --dry-run",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -287,7 +288,7 @@ fn test_cmd_delete_index_dry_run() -> Result<()> {
     // Non-empty index
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {} --dry-run",
+            "index delete {} --metastore-uri {} --dry-run",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -316,7 +317,7 @@ async fn test_cmd_delete() -> Result<()> {
     );
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {}",
+            "index gc {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -329,7 +330,7 @@ async fn test_cmd_delete() -> Result<()> {
 
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {}",
+            "index delete {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -363,7 +364,7 @@ async fn test_cmd_garbage_collect_no_grace() -> Result<()> {
     assert_eq!(splits.len(), 1);
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {}",
+            "index gc {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -383,7 +384,7 @@ async fn test_cmd_garbage_collect_no_grace() -> Result<()> {
         .await?;
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {} --dry-run --grace-period 10m",
+            "index gc {} --metastore-uri {} --dry-run --grace-period 10m",
             index_id, test_env.metastore_uri,
         )
         .as_str(),
@@ -403,7 +404,7 @@ async fn test_cmd_garbage_collect_no_grace() -> Result<()> {
 
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {} --grace-period 10m",
+            "index gc {} --metastore-uri {} --grace-period 10m",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -428,7 +429,7 @@ async fn test_cmd_garbage_collect_no_grace() -> Result<()> {
 
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {}",
+            "index delete {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -456,7 +457,7 @@ async fn test_cmd_garbage_collect_spares_files_within_grace_period() -> Result<(
     assert_eq!(splits.len(), 1);
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {}",
+            "index gc {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -485,7 +486,7 @@ async fn test_cmd_garbage_collect_spares_files_within_grace_period() -> Result<(
 
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {} --grace-period 2s",
+            "index gc {} --metastore-uri {} --grace-period 2s",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -502,7 +503,7 @@ async fn test_cmd_garbage_collect_spares_files_within_grace_period() -> Result<(
     sleep(Duration::from_secs(3)).await;
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {} --dry-run --grace-period 2s",
+            "index gc {} --metastore-uri {} --dry-run --grace-period 2s",
             index_id, test_env.metastore_uri,
         )
         .as_str(),
@@ -517,7 +518,7 @@ async fn test_cmd_garbage_collect_spares_files_within_grace_period() -> Result<(
 
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {} --grace-period 2s",
+            "index gc {} --metastore-uri {} --grace-period 2s",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -541,9 +542,9 @@ async fn test_cmd_dry_run_delete_on_s3_localstack() -> Result<()> {
     let index_uri = test_env.index_uri(&index_id);
     make_command(
         format!(
-            "new --index-uri {} --index-id {} --metastore-uri {} --index-config-path {}",
-            index_uri,
+            "index create {} --index-uri {} --metastore-uri {} --index-config-path {}",
             index_id,
+            index_uri,
             test_env.metastore_uri,
             test_env.resource_files["config"].display()
         )
@@ -561,7 +562,7 @@ async fn test_cmd_dry_run_delete_on_s3_localstack() -> Result<()> {
 
     make_command(
         format!(
-            "gc --index-id {} --metastore-uri {}",
+            "index gc {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -574,7 +575,7 @@ async fn test_cmd_dry_run_delete_on_s3_localstack() -> Result<()> {
 
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {} --dry-run",
+            "index delete {} --metastore-uri {} --dry-run",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -588,7 +589,7 @@ async fn test_cmd_dry_run_delete_on_s3_localstack() -> Result<()> {
 
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {}",
+            "index delete {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -607,9 +608,9 @@ async fn test_all_local_index() -> Result<()> {
     let test_env = create_test_env(TestStorageType::LocalFileSystem)?;
     make_command(
         format!(
-            "new --index-uri {} --index-id {} --metastore-uri {} --index-config-path {}",
-            test_env.index_uri(&index_id),
+            "index create {} --index-uri {} --metastore-uri {} --index-config-path {}",
             index_id,
+            test_env.index_uri(&index_id),
             test_env.metastore_uri,
             test_env.resource_files["config"].display()
         )
@@ -635,7 +636,7 @@ async fn test_all_local_index() -> Result<()> {
     // TODO: randomize port.
     let mut server_process = spawn_command(
         format!(
-            "serve --metastore-uri {} --host 127.0.0.1 --port 8182",
+            "service start searcher --metastore-uri {} --host 127.0.0.1 --port 8182",
             test_env.metastore_uri,
         )
         .as_str(),
@@ -678,7 +679,7 @@ async fn test_all_local_index() -> Result<()> {
 
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {}",
+            "index delete {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -705,9 +706,9 @@ async fn test_all_with_s3_localstack_cli() -> Result<()> {
 
     make_command(
         format!(
-            "new --index-uri {} --index-id {} --metastore-uri {} --index-config-path {}",
-            index_uri,
+            "index create {} --index-uri {} --metastore-uri {} --index-config-path {}",
             index_id,
+            index_uri,
             test_env.metastore_uri,
             test_env.resource_files["config"].display()
         )
@@ -729,7 +730,7 @@ async fn test_all_with_s3_localstack_cli() -> Result<()> {
     // cli search
     make_command(
         format!(
-            "search --index-id {} --metastore-uri {} --query level:info",
+            "index search {} --metastore-uri {} --query level:info",
             index_id, test_env.metastore_uri,
         )
         .as_str(),
@@ -745,7 +746,7 @@ async fn test_all_with_s3_localstack_cli() -> Result<()> {
     // TODO: ditto.
     let mut server_process = spawn_command(
         format!(
-            "serve --metastore-uri {} --host 127.0.0.1 --port 8182",
+            "service start searcher --metastore-uri {} --host 127.0.0.1 --port 8182",
             test_env.metastore_uri,
         )
         .as_str(),
@@ -777,7 +778,7 @@ async fn test_all_with_s3_localstack_cli() -> Result<()> {
 
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {}",
+            "index delete {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
@@ -817,7 +818,7 @@ async fn test_all_with_s3_localstack_internal_api() -> Result<()> {
     // cli search
     make_command(
         format!(
-            "search --index-id {} --metastore-uri {} --query level:info",
+            "index search {} --metastore-uri {} --query level:info",
             index_id, test_env.metastore_uri,
         )
         .as_str(),
@@ -833,7 +834,7 @@ async fn test_all_with_s3_localstack_internal_api() -> Result<()> {
     // TODO: ditto.
     let mut server_process = spawn_command(
         format!(
-            "serve --metastore-uri {} --host 127.0.0.1 --port 8182",
+            "service start searcher --metastore-uri {} --host 127.0.0.1 --port 8182",
             test_env.metastore_uri,
         )
         .as_str(),
@@ -865,7 +866,7 @@ async fn test_all_with_s3_localstack_internal_api() -> Result<()> {
 
     make_command(
         format!(
-            "delete --index-id {} --metastore-uri {}",
+            "index delete {} --metastore-uri {}",
             index_id, test_env.metastore_uri
         )
         .as_str(),
